chore: add all contracts option
