+++ frontend/src/app/(routes)/cosmwasm/components/Contracts.tsx
@@ -1,9 +1,15 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import SearchContracts from './SearchContracts';
 import DeployContract from './DeployContract';
 import ContractInfo from './ContractInfo';
+import { useSearchParams } from 'next/navigation';
+import useContracts from '@/custom-hooks/useContracts';
+import useGetChainInfo from '@/custom-hooks/useGetChainInfo';
+import { setContract } from '@/store/features/cosmwasm/cosmwasmSlice';
+import { useAppDispatch } from '@/custom-hooks/StateHooks';
 
 const Contracts = ({ chainID }: { chainID: string }) => {
+  const dispatch = useAppDispatch();
   const [deployContractOpen, setDeployContractOpen] = useState(false);
   const [selectedContract, setSelectedContract] = useState({
     address: '',
@@ -13,6 +19,41 @@ const Contracts = ({ chainID }: { chainID: string }) => {
     setDeployContractOpen(false);
     setSelectedContract({ address, name });
   };
+
+  const { getChainInfo } = useGetChainInfo();
+  const { restURLs } = getChainInfo(chainID);
+
+  const { getContractInfo } = useContracts();
+
+  const paramsContractAddress = useSearchParams().get('contract');
+
+  const fetchContractInfo = async () => {
+    const { data } = await getContractInfo({
+      address: paramsContractAddress || '',
+      baseURLs: restURLs,
+    });
+    if (data) {
+      dispatch(
+        setContract({
+          chainID,
+          contractAddress: data?.address,
+          contractInfo: data?.contract_info,
+        })
+      );
+      setSelectedContract({
+        address: data?.address,
+        name: data?.contract_info?.label,
+      });
+    }
+  };
+
+  useEffect(() => {
+    console.log('herer111');
+    if (paramsContractAddress?.length) {
+      console.log('herer');
+      fetchContractInfo();
+    }
+  }, [paramsContractAddress]);
 
   return (
     <div className="h-full flex flex-col gap-10">
